# api/docs/telephant-api-documentation.yml
openapi: 3.0.0
info:
  title: Telephant API
  description: |
    *The mechanism for swapping airtime is powered by Telephant's APIs. Developers can use the API to tap into the platform's capabilities and incorporate the service into their own apps.*

    The Telephant API's backbone is HTTP. The nature of our API is RESTful. It has a set of predictable resource URLs. It returns HTTP response codes to indicate errors. It also accepts JSON as an HTTP body payload and returns it, along with some other useful formats.
    
    You can use your favorite HTTP/REST library for your programming language or an HTTP client like Postman or CURL to use Telephant's API.

    Happy programming **:-)**

    ## HTTP Response Status Codes
    HTTP response status codes show whether or not a particular HTTP request was completed successfully. Section 10 of RFC 2616 defines the status codes listed below. RFC 7231 contains an updated specification.

    Telephant API makes use of the following HTTP status codes, grouped into three classes:

    ### Successful Responses - 2XX

    #### 200 - OK

    The request has succeeded.

    #### 201 - Created

    As a result of the request's success, a new resource has been generated. This is the response that is normally sent in response to POST or some PUT queries.

    ### Client Error - 4XX
    
    #### 400 - Bad Request
    Due to incorrect syntax, the server was unable to understand the request.
    
    #### 401 - Unauthorized
...
    
...
    
...
    
...

...

...

...

...

...
servers:
...
paths:
...
  /v1/notifications:
    get:
      summary: List of notifications.
      description: Get the list of notifications for an individual or group account.
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
        - in: query
          name: page
          description: The page number of results to return.
          schema:
            type: integer
        - in: query
          name: limit
          description: The numbers of items to return per page.
          schema:
            type: integer
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              examples:
                Notifcations-Get:
                  $ref: '#/components/examples/Notifications-Get'

        '4XX':
          $ref: '#/components/responses/4XX'
        '5XX':
          $ref: '#/components/responses/5XX'

    post:
      summary: Mark all notifications read.
      description: Changes the status of all notifications to read.
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              example:
                {
                    message: '10 notifications were marked as read.'
                }
        '4XX':
          $ref: '#/components/responses/4XX'
        '5XX':
          $ref: '#/components/responses/5XX'
...
  /v1/notifications/{NotificationID}:
    get:
      summary: Get specific notification.
      description: Retrieve the details of a specific notification by it's id.
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
        - in: path
          name: NotificationID
          description: Use a valid resource id to identify the notification to query.
          schema:
            type: string
            format: uuid
          required: true
          example: '60ebd53e2f79311541619554'
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              examples:
                Notifications-Get-Item:
                  $ref: '#/components/examples/Notifications-Get-Item'
        '4XX':
          $ref: '#/components/responses/4XX'
        '5XX':
          $ref: '#/components/responses/5XX'

    post:
      summary:  Mark notification as read.
      description: Mark a specific notification as read for an individual or group account.
      tags:
        - Notifications
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
        - in: path
          name: NotificationID
          description: Use a valid resource id to identify the notification to query.
          schema:
            type: string
            format: uuid
          required: true
          example: '60ebd53e2f79311541619554'
      responses:
        '2XX':
          description: Success
          content:
            application/json:
              example:
                {
                    message: 'The notification was marked as read.'
                }
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

  #### PING ####
...
  /v1/retailers:
    get:
      summary: List retailers.
      description: Get a listing of Telephant retailers.
      tags:
        - Retailers
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
    post:
      summary: Create a retailer listing.
      description: Create a Telephant retailer listing.
      tags:
        - Retailers
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
  /v1/retailers/{RetailerID}:
    put:
      summary: Update retailer.
      description: Update a retailer listing.
      tags:
        - Retailers
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: RetailerID
          schema:
            type: string
            format: uuid
          required: true
          description: Unique ID of the retailer listing.
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
    delete:
      summary: Delete retailer.
      description: Delete a retailer listing.
      tags:
        - Retailers
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: RetailerID
          schema:
            type: string
            format: uuid
          required: true
          description: Unique ID of the retailer listing.
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

#### TRANSACTIONS ####
...
  /v1/users?action=signup:
    post:
      summary: User signup.
      description: An end point to sign up as a user.
      tags:
        - Users
      parameters:
        - in: query
          name: action
          schema:
            type: string
            enum:
              - signup
          description: |

            User actions can be:
             * `signup` - Create a user account.
             * `login` - As a user, sign in.
      requestBody:
        description: Provide the telephone number and a password to signup.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  
                  message: 
                    type: string
                    example: 'The user was successfully registered.'
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                Password-complexity:
                  value:
                    {
                      "code": 400,
                      "name": "BadRequestError",
                      "message": 'User validation failed : Password not complex enough.'
                    }
                South-Africa-format:
                  value:
                    {
                      "code": 400,
                      "name": "BadRequestError",
                      "message": 'User validation failed : The telephone number does not conform to the South African format (27XXXXXXXXX).'
                    }
                User-taken:
                  value:
                    {
                      "code": 400,
                      "name": "BadRequestError",
                      "message": "This telephone number is already registered."
                    }
        '5xx':
          $ref: '#/components/responses/5XX'
  /v1/users?action=login:
    post:
      summary: User login.
      description: An end point to log in a user.
      tags:
        - Users
      parameters:
        - in: query
          name: action
          schema:
            type: string
            enum:
              - login
          description: |

            User actions can be:
             * `signup` - Create a user account.
             * `login` - As a user, sign in.
      requestBody:
        description: Provide the telephone number and a password to login.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User successfully logged in.'
                  token:
                    type: string
                    example: 'BEARER <TOKEN STRING>'
        '401':
          description: Unauthorised
          content:
            application/json:
              example:
                  {
                    "code": 401,
                    "name": "UnauthorizedError",
                    "message": "Unable to authenticate using credentials provided."
                  }
        '5xx':
          $ref: '#/components/responses/5XX'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Users:
      type: object
      properties:
        telephoneNumber:
          type: string
          pattern: '^(27)[0-9]{9}$'
          description: Provide a valid South African mobile telphone number in the format 27XXXXXXXXX.
          example: "27821234567"
        password:
          type: string
          pattern: '(?=^.{8,}$)(?=.*\d)(?=.*[!@#$%^&*]+)(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$'
          description: |
            Provide a suitable password.
              * The password length must be greater than or equal to 8
              * The password must contain one or more uppercase characters
              * The password must contain one or more lowercase characters
              * The password must contain one or more numeric values
              * The password must contain one or more special characters
            example: P@ssword123
...
  responses:
    4XX:
      description: |
        Client error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 400
                description: |
                  Please review the documentation for [HTTP Response Status Codes](/#section/HTTP-Response-Status-Codes)

                  Possible values are:
                    - 400 - Bad Request
                    - 401 - Unauthorized
                    - 403 - Forbidden
                    - 404 - Not Found
                    - 405 - Method Not Allowed
              name:
                type: String
                example: 'BadRequest'
                description: This value will always be returned as 'bumpy case'.
              message:
                type: String
                description: |
                  The message field will provide specific information as to the reason for the error.
                example: 'The account ID provided is not valid.'
    5XX:
      description: |
        Server error  
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 500
              name:
                type: String
                example: 'InternalServerError'
              message:
                type: String
                example: 'We are sorry but something went wrong on our side.'
  examples:
    Notifications-Get-Item:
      value:
        {
            "dateTime": "2021-05-23T07:18:00.000Z",
            "read": false,
            "type": "INFO",
            "_id": "60fed5f105b740691acc2d91",
            "subject": "Sent",
            "message": "You sent Napoleon R 50",
            "owner": "my-business-1",
            "link": "y3huwi17",
            "createdAt": "2021-07-26T15:34:09.689Z",
            "updatedAt": "2021-07-26T15:34:09.689Z",
            "__v": 0
        }
...
    Notifications-Get:
      value:
        {
          "page": 1,
          "total_pages": 1,
          "total_results": 2,
          "results": [
                        {
                            "dateTime": "2021-05-11T16:55:00.000Z",
                            "read": false,
                            "type": "INFO",
                            "_id": "60fed5f105b740691acc2d8d",
                            "subject": "Top Up",
                            "message": "You deposited R 1500 into your telphant account.",
                            "owner": "my-business-1",
                            "link": "3zm9e5o5",
                            "createdAt": "2021-07-26T15:34:09.689Z",
                            "updatedAt": "2021-07-26T15:34:09.689Z",
                            "__v": 0
                        },
                        {
                            "dateTime": "2021-05-23T07:18:00.000Z",
                            "read": false,
                            "type": "INFO",
                            "_id": "60fed5f105b740691acc2d91",
                            "subject": "Sent",
                            "message": "You sent Napoleon R 50",
                            "owner": "my-business-1",
                            "link": "y3huwi17",
                            "createdAt": "2021-07-26T15:34:09.689Z",
                            "updatedAt": "2021-07-26T15:34:09.689Z",
                            "__v": 0
                        }
                      ]
        }
...